web-dashboard/api/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
anthropic>=0.28.0
python-dotenv>=1.0.0
redis>=5.0.0
